+++ MovieQuiz/Models/QuiszStepViewModel.swift
@@ -1,8 +1,7 @@
-//
-//  QuiszstepViewModel.swift
-//  MovieQuiz
-//
-//  Created by  Admin on 24.01.2024.
-//
+import UIKit
 
-import Foundation
+struct QuizStepViewModel {
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
