+++ MovieQuiz/Models/QuizQuestion.swift
@@ -6,3 +6,9 @@
 //
 
 import Foundation
+
+struct QuizQuestion {
+    let image: String
+    let text: String
+    let correctAnswer: Bool
+}
