+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -6,3 +6,9 @@
 //
 
 import Foundation
+
+struct QuizResultsViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
